requirements.txt
pypose >= 0.6.8
opencv-python-headless
evo
matplotlib==3.8.4
tabulate
tqdm
rich
cupy-cuda12x
einops
timm==0.9.12
rerun-sdk==0.17
yacs
# torch
numpy
pyyaml
wandb
# torchvision
pillow
scipy
flow_vis
h5py
xformers==0.0.27.post2
# tensorrt
onnx
